+++ src/Spd.Presentation.Screening/ClientApp/src/app/core/code-types/code-types.models.ts
@@ -2,3 +2,4 @@ export * from 'src/app/api/models';
 export { ApplicationOriginTypeCode } from './application-origin-type.model';
 export { ApplicationPortalStatisticsTypeCode } from './application-portal-statistics-type.model';
 export { GenderTypeCode } from './gender-type.model';
+export { PortalTypeCode } from './portal-type.model';
