+++ src/Spd.Presentation.Screening/ClientApp/src/app/core/code-types/portal-type.model.ts
@@ -1,0 +1,4 @@
+export enum PortalTypeCode {
+	Crrp = 'CRRP',
+	Psso = 'PSSO',
+}
