+++ src/Spd.Presentation.Screening/ClientApp/src/app/modules/crrp-portal/components/application-statuses.component.ts
@@ -1,5 +1,6 @@
 import { Component } from '@angular/core';
 import { Router } from '@angular/router';
+import { PortalTypeCode } from 'src/app/core/code-types/portal-type.model';
 import { ScreeningStatusResponse } from 'src/app/shared/components/screening-statuses-common.component';
 import { CrrpRoutes } from '../crrp-routing.module';
 
@@ -8,25 +9,18 @@ import { CrrpRoutes } from '../crrp-routing.module';
 	template: `
 		<app-crrp-header></app-crrp-header>
 
-		<section class="step-section my-3 px-md-4 py-md-3 p-sm-0">
-			<div class="row">
-				<div class="col-xl-8 col-lg-10 col-md-12 col-sm-12">
-					<h2 class="mb-2 fw-normal">Application Statuses</h2>
-					<app-applications-banner></app-applications-banner>
-				</div>
-			</div>
-
-			<app-screening-statuses-common
-				portal="CRRP"
-				title="Application Statuses"
-				(emitPayNow)="onPayNow($event)"
-				(emitVerifyIdentity)="onVerifyIdentity($event)"
-			></app-screening-statuses-common>
-		</section>
+		<app-screening-statuses-common
+			[portal]="portal.Crrp"
+			title="Application Statuses"
+			(emitPayNow)="onPayNow($event)"
+			(emitVerifyIdentity)="onVerifyIdentity($event)"
+		></app-screening-statuses-common>
 	`,
 	styles: [],
 })
 export class ApplicationStatusesComponent {
+	portal = PortalTypeCode;
+
 	constructor(private router: Router) {}
 
 	onPayNow(application: ScreeningStatusResponse): void {
