+++ src/Spd.Presentation.Screening/ClientApp/src/app/modules/psso-portal/components/identity-verification.component.ts
@@ -1,15 +1,11 @@
 import { Component } from '@angular/core';
+import { PortalTypeCode } from 'src/app/core/code-types/portal-type.model';
 
 @Component({
-  selector: 'app-identity-verification',
-  template: `
-    <p>
-      identity-verification works!
-    </p>
-  `,
-  styles: [
-  ]
+	selector: 'app-identity-verification',
+	template: ` <app-identify-verification-common [portal]="portal.Psso"></app-identify-verification-common> `,
+	styles: [],
 })
 export class IdentityVerificationComponent {
-
+	portal = PortalTypeCode;
 }
