+++ src/Spd.Presentation.Screening/ClientApp/src/app/modules/psso-portal/components/manual-submission.component.ts
@@ -1,15 +1,11 @@
 import { Component } from '@angular/core';
+import { PortalTypeCode } from 'src/app/core/code-types/portal-type.model';
 
 @Component({
-  selector: 'app-manual-submission',
-  template: `
-    <p>
-      manual-submission works!
-    </p>
-  `,
-  styles: [
-  ]
+	selector: 'app-manual-submission',
+	template: ` <app-manual-submission-common [portal]="portal.Psso"></app-manual-submission-common> `,
+	styles: [],
 })
 export class ManualSubmissionComponent {
-
+	portal = PortalTypeCode;
 }
