+++ src/Spd.Presentation.Screening/ClientApp/src/app/modules/psso-portal/psso-routing.module.ts
@@ -27,7 +27,7 @@ const routes: Routes = [
 			{ path: PssoRoutes.SCREENING_STATUSES, component: ScreeningStatusesComponent },
 			{ path: PssoRoutes.IDENTITY_VERIFICATION, component: IdentityVerificationComponent },
 			{ path: PssoRoutes.MANUAL_SUBMISSIONS, component: ManualSubmissionComponent },
-			{ path: '', redirectTo: PssoRoutes.SCREENING_STATUSES, pathMatch: 'full' },
+			{ path: '', redirectTo: PssoRoutes.path(PssoRoutes.SCREENING_STATUSES), pathMatch: 'full' },
 		],
 	},
 ];
